only react to autocommands from own vim server and implement :q

+++ pida/editors/vim/vim.py
@@ -176,26 +176,34 @@ class VimCallback(object):
             self.svc.init_vim_server()
 
     def vim_bufferchange(self, server, cwd, file_name, bufnum):
-        if file_name:
-            if os.path.abspath(file_name) != file_name:
-                file_name = os.path.join(cwd, file_name)
-            if os.path.isdir(file_name):
-                self.svc.boss.cmd('filemanager', 'browse', new_path=file_name)
-                self.svc.boss.cmd('filemanager', 'present_view')
-                self.svc.open_last()
-            else:
-                self.svc.boss.cmd('buffer', 'open_file', file_name=file_name)
+        if server == self.svc.server:
+            if file_name:
+                if os.path.abspath(file_name) != file_name:
+                    file_name = os.path.join(cwd, file_name)
+                if os.path.isdir(file_name):
+                    self.svc.boss.cmd('filemanager', 'browse', new_path=file_name)
+                    self.svc.boss.cmd('filemanager', 'present_view')
+                    self.svc.open_last()
+                else:
+                    self.svc.boss.cmd('buffer', 'open_file', file_name=file_name)
 
     def vim_bufferunload(self, server, file_name):
-        if file_name:
-            self.svc.remove_file(file_name)
-            self.svc.boss.get_service('buffer').cmd('close_file', file_name=file_name)
+        if server == self.svc.server:
+            if file_name:
+                self.svc.remove_file(file_name)
+                self.svc.boss.get_service('buffer').cmd('close_file', file_name=file_name)
 
     def vim_filesave(self, server, file_name):
-        self.svc.boss.cmd('buffer', 'current_file_saved')
+        if server == self.svc.server:
+            self.svc.boss.cmd('buffer', 'current_file_saved')
 
     def vim_cursor_move(self, server, line_number):
-        self.svc.set_current_line(int(line_number))
+        if server == self.svc.server:
+            self.svc.set_current_line(int(line_number))
+
+    def vim_shutdown(self, server, args):
+        if server == self.svc.server:
+            self.svc.boss.stop(force=True)
 
 
 # Service class
